requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-jose
passlib[bcrypt]
python-dotenv
cryptography